Remove the `fields` attribute of the `dv-widget` tag
If a widget needs to provide field values to the widget it includes it can do so
using the `init` attribute. The widget fields will be initialized to the values
provided in `init`.

Cliché widgets can no longer pass arbitrary information to an included widget.
Information is either passed through bonds or through the `init` attribute.