allow the user to specify the parent of an action with css classes
At runtime, a child action might end up in a different position of the tree than
the one specified by the HTML of its parent action. This commit allows the
author of a cliche to attach a special class to an action to point to its
parent. This information will be used at runtime to find the action parent
instead of relying on its position in the DOM